backend_view_wise/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
dataclasses-json==0.6.7
deepdiff==8.2.0
defusedxml==0.7.1
Django==5.1.6
django-allauth==65.4.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-yasg==1.21.8
frozenlist==1.5.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
inflection==0.5.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langchain-unstructured==0.1.5
langsmith==0.3.8
Markdown==3.7
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.3
oauthlib==3.2.2
orderly-set==5.3.0
orjson==3.10.15
packaging==24.2
propcache==0.2.1
psycopg==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.9.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
tenacity==9.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unstructured-client==0.25.9
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
youtube-dl==2021.12.17
zstandard==0.23.0
